+++ lib/slugify/__init__.py
@@ -1,0 +1,4 @@
+#v1.2.4
+#https://github.com/un33k/python-slugify/releases
+
+from slugify import *